Create Card form shows up at create/card. Moved static back into core
