+++ core/forms.py
@@ -5,4 +5,4 @@ class CreateCardForm(forms.ModelForm):
 
     class Meta:
         model = Card
-        fields = ('question', 'answer')
+        fields = ('question', 'answer',)
