+++ core/static/css/main.css
@@ -1,0 +1,3 @@
+header {
+    color: red;
+}
