+++ core/static/js/main.js
@@ -1,4 +1,7 @@
-// functions that create shortcuts:
+/* globals fetch */
+const Cookies = require('js-cookie')
+const deepmerge = require('deepmerge')
+
 function qS (selector) {
   return document.querySelector(selector)
 }
@@ -7,8 +10,14 @@ function qSA (selector) {
   return document.querySelectorAll(selector)
 }
 
-function createCard (event) {
-  const createCardDiv = qs('#create-card')
+// This function will be helpful if we need to fetch anything from an API, for example, fetching trivia questions from this API: https://opentdb.com/api_config.php
+function request (url, options) {
+  if (!options) {
+    options = {}
+  }
+  const defaultOptions = {
+    headers: { 'X-CSRFToken': Cookies.get('csrftoken'), 'X-Requested-With': 'XMLHttpRequest' }
+  }
 
-  createCardDiv.innerHTML = ''
+  return fetch(url, deepmerge(defaultOptions, options))
 }
