+++ core/templates/base.html
@@ -14,7 +14,7 @@
         <!-- From browserify -->
         <script src="{% static 'build/bundle.js' %}"></script>
         <!-- Add additional CSS in static file -->
-        <!-- <link rel="stylesheet" href="{% static 'css/styles.css' %}"> -->
+        <link rel="stylesheet" href="{% static 'css/main.css' %}">
     </head>
     <body>
         <header>
