+++ core/templates/create_card.html
@@ -1,0 +1,12 @@
+{% extends "base.html" %}
+
+{% block content %}
+
+    <h2 class="">Create a card.</h2>
+        <form method='POST'>
+            {% csrf_token %}
+                {{ form.as_p }}
+                <button type="submit" class="create_card">Create</button>
+        </form>
+
+{% endblock %}
