+++ core/urls.py
@@ -9,5 +9,5 @@ urlpatterns = [
     path('create/', views.create, name='create'),
     path('play/', views.play, name='play'),
     path('deck_detail/<slug:slug>', views.deck_detail, name='deck_detail'),
-    path('deck/<slug:slug>/create/', views.create_card, name='create_card'),
+    path('create/card', views.create_card, name='create_card'),
 ]
