--- static/js/main.js
@@ -1,23 +1,0 @@
-/* globals fetch */
-const Cookies = require('js-cookie')
-const deepmerge = require('deepmerge')
-
-function qS (selector) {
-  return document.querySelector(selector)
-}
-
-function qSA (selector) {
-  return document.querySelectorAll(selector)
-}
-
-// This function will be helpful if we need to fetch anything from an API, for example, fetching trivia questions from this API: https://opentdb.com/api_config.php
-function request (url, options) {
-  if (!options) {
-    options = {}
-  }
-  const defaultOptions = {
-    headers: { 'X-CSRFToken': Cookies.get('csrftoken'), 'X-Requested-With': 'XMLHttpRequest' }
-  }
-
-  return fetch(url, deepmerge(defaultOptions, options))
-}
